Separate user provider into wallet and user
