+++ packages/atlas/src/providers/user/user.hooks.ts
@@ -1,36 +1,24 @@
-import { useSignerWallet } from './user.helpers'
 import { useUserContext } from './user.provider'
 import { UserStoreActions, useUserStore } from './user.store'
-import { ActiveUserState, SignerWalletAccount, SignerWalletStatus, UserContextValue } from './user.types'
+import { ActiveUserState, UserContextValue } from './user.types'
 
 type UseUserHookReturn = ActiveUserState &
   UserStoreActions &
   UserContextValue & {
-    walletAccounts: SignerWalletAccount[]
-    walletStatus: SignerWalletStatus
-
     isLoggedIn: boolean
-    isWalletLoading: boolean
-
-    getWalletsList: ReturnType<typeof useSignerWallet>['getWalletsList']
   }
 
 export const useUser = (): UseUserHookReturn => {
-  const { accountId, memberId, channelId, walletStatus, walletAccounts, actions } = useUserStore()
+  const { accountId, memberId, channelId, actions } = useUserStore()
   const userContext = useUserContext()
-  const { getWalletsList } = useSignerWallet()
 
-  const isLoggedIn = !!accountId && !!memberId && walletStatus === 'connected'
+  const isLoggedIn = !!accountId && !!memberId
 
   return {
     accountId,
     memberId,
     channelId,
-    walletStatus,
-    walletAccounts,
     isLoggedIn,
-    isWalletLoading: walletStatus === 'pending',
-    getWalletsList,
     ...userContext,
     ...actions,
   }
