+++ packages/atlas/src/providers/user/user.store.ts
@@ -1,14 +1,9 @@
 import { createStore } from '@/utils/store'
 
-import { ActiveUserState, SignerWallet, SignerWalletAccount, SignerWalletStatus } from './user.types'
+import { ActiveUserState } from './user.types'
 
 export type UserStoreState = ActiveUserState & {
   userId: string | null
-  wallet: SignerWallet | null
-  walletAccounts: SignerWalletAccount[]
-  walletStatus: SignerWalletStatus
-  lastUsedWalletName: string | null
-  lastChainMetadataVersion: [string, number] | null // [genesisHash, number]
   signInModalOpen: boolean
 }
 
@@ -16,14 +11,8 @@ export type UserStoreActions = {
   resetActiveUser: () => void
   setActiveUser: (activeUserChanges: Partial<ActiveUserState>) => void
   signOut: () => void
-
   setUserId: (userId: string | null) => void
-  setWallet: (wallet: SignerWallet) => void
-  setWalletAccounts: (accounts: SignerWalletAccount[]) => void
-  setWalletStatus: (status: SignerWalletStatus) => void
-
   setSignInModalOpen: (isOpen: boolean) => void
-  setLastChainMetadataVersion: (genesisHash: string, version: number) => void
 }
 
 export const useUserStore = createStore<UserStoreState, UserStoreActions>(
@@ -34,11 +23,6 @@ export const useUserStore = createStore<UserStoreState, UserStoreActions>(
       channelId: null,
 
       userId: null,
-      wallet: null,
-      walletAccounts: [],
-      walletStatus: 'unknown',
-      lastUsedWalletName: null,
-      lastChainMetadataVersion: null,
       signInModalOpen: false,
     },
     actionsFactory: (set) => ({
@@ -61,10 +45,6 @@ export const useUserStore = createStore<UserStoreState, UserStoreActions>(
           state.accountId = null
           state.memberId = null
           state.channelId = null
-          state.wallet = null
-          state.walletStatus = 'unknown'
-          state.walletAccounts = []
-          state.lastUsedWalletName = null
         })
       },
 
@@ -73,30 +53,9 @@ export const useUserStore = createStore<UserStoreState, UserStoreActions>(
           state.userId = userId
         })
       },
-      setWallet: (wallet) => {
-        set((state) => {
-          state.wallet = wallet
-          state.lastUsedWalletName = wallet.extensionName
-        })
-      },
-      setWalletAccounts: (accounts) => {
-        set((state) => {
-          state.walletAccounts = accounts
-        })
-      },
-      setWalletStatus: (status) => {
-        set((state) => {
-          state.walletStatus = status
-        })
-      },
       setSignInModalOpen: (isOpen) => {
         set((state) => {
           state.signInModalOpen = isOpen
-        })
-      },
-      setLastChainMetadataVersion: (genesisHash, version) => {
-        set((state) => {
-          state.lastChainMetadataVersion = [genesisHash, version]
         })
       },
     }),
@@ -105,7 +64,7 @@ export const useUserStore = createStore<UserStoreState, UserStoreActions>(
     persist: {
       key: 'activeUser',
       version: 0,
-      whitelist: ['accountId', 'memberId', 'channelId', 'lastUsedWalletName', 'lastChainMetadataVersion', 'userId'],
+      whitelist: ['accountId', 'memberId', 'channelId', 'userId'],
       migrate: (oldState) => {
         return oldState
       },
