+++ packages/atlas/src/providers/user/user.types.ts
@@ -1,5 +1,3 @@
-import { Wallet, WalletAccount } from '@talismn/connect-wallets'
-
 import { useMemberships } from '@/api/hooks/membership'
 import { FullMembershipFieldsFragment } from '@/api/queries/__generated__/fragments.generated'
 
@@ -10,18 +8,10 @@ export type ActiveUserState = {
   channelId: string | null
 }
 
-export type SignerWallet = Wallet
-export type SignerWalletStatus = 'unknown' | 'connected' | 'disconnected' | 'pending'
-export type SignerWalletAccount = WalletAccount
 export type UserContextValue = {
   memberships: Membership[]
   membershipsLoading: boolean
   activeMembership: Membership | null
-  isSignerMetadataOutdated: boolean
   activeChannel: Membership['channels'][number] | null
-  isAuthLoading: boolean
-  signIn: (walletName?: string, mobileCallback?: ({ onConfirm }: { onConfirm: () => void }) => void) => Promise<boolean>
-  updateSignerMetadata: () => Promise<boolean>
-  skipSignerMetadataUpdate: () => Promise<void>
   refetchUserMemberships: ReturnType<typeof useMemberships>['refetch']
 }
