+++ packages/atlas/src/providers/wallet/wallet.hooks.ts
@@ -1,0 +1,14 @@
+import { useWalletContext } from '@/providers/wallet/wallet.provider'
+import { useWalletStore } from '@/providers/wallet/wallet.store'
+
+export const useWallet = () => {
+  const { walletStatus, walletAccounts, wallet } = useWalletStore()
+  const walletContext = useWalletContext()
+
+  return {
+    walletStatus,
+    walletAccounts,
+    wallet,
+    ...walletContext,
+  }
+}
